% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityParameter_onefish.R
\name{activityCompressibility_onefish}
\alias{activityCompressibility_onefish}
\title{Title}
\usage{
activityCompressibility_onefish(ffc, zhc, bin_nsecs, fps)
}
\arguments{
\item{fps}{}
}
\description{
Title
}
