% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityParameter_onefish.R
\name{activityPercentageTimeActive_onefish}
\alias{activityPercentageTimeActive_onefish}
\title{Title}
\usage{
activityPercentageTimeActive_onefish(ffc, zhc, bin_nsecs, fps)
}
\arguments{
\item{fps}{}
}
\description{
Title
}
