% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTimetoRAWs.R
\name{addTimetoRAWs}
\alias{addTimetoRAWs}
\title{Imports RAWs.csv}
\usage{
addTimetoRAWs(ffpath, zebpath, dayduration = 14)
}
\arguments{
\item{ffpath}{Path to frame-by-frame data, typically ..._RAWs.csv.}

\item{zebpath}{Path to Zebralab file, typically YYMMDD_BX.xls.}

\item{dayduration}{Duration of day in hours. Default: 14.}
}
\value{
Dataframe containing frame-by-frame data with added timestamp columns.
}
\description{
Imports RAWs.csv
}
