% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustPixel.R
\name{adjustPixel}
\alias{adjustPixel}
\title{Title}
\usage{
adjustPixel(
  ffpath,
  genopath,
  grpL,
  grpS = NA,
  scale = NA,
  round = "down",
  dayduration = 14
)
}
\arguments{
\item{round}{}
}
\description{
Title
}
