% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allparameters}
\alias{allparameters}
\title{All Parameters}
\format{
\subsection{\verb{[1] "activityPercentageTimeActive" "activityTotalPx"              "activitySunsetStartle"        "activitySlope"}}{

A vector of strings:
\describe{
\item{allparameters}{All parameters}
}
}
}
\usage{
allparameters
}
\description{
Simple vector with all parameter names, e.g. activeboutMean.
}
\keyword{datasets}
