% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alluparams}
\alias{alluparams}
\title{All Unique Parameters}
\format{
\subsection{\code{alluparams}}{

A vector of strings:
\describe{
\item{alluparams}{All unique parameters}
}
}
}
\usage{
alluparams
}
\description{
Simple vector with all unique parameter names, e.g. day_activeboutMean.
}
\keyword{datasets}
