% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{appendRAWs}
\alias{appendRAWs}
\title{Title}
\usage{
appendRAWs(
  ffpaths,
  zebpaths,
  zebDeprecatedFormat = FALSE,
  dayduration = 14,
  exportPath
)
}
\arguments{
\item{exportPath}{}
}
\description{
Title
}
