% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAddTimetoRAWs.R
\name{averageFramerate}
\alias{averageFramerate}
\title{Calculates average framerate}
\usage{
averageFramerate(tsecs)
}
\arguments{
\item{tsecs}{Timestamps in seconds}
}
\value{
Average framerate
}
\description{
Calculates average framerate from a series of timestamps in seconds.
}
\examples{
averageFramerate(c(0, 0.04, 0.08))
## returns 25 frames-per-second
}
