% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{calculateFingerprint}
\alias{calculateFingerprint}
\title{Title}
\usage{
calculateFingerprint(
  paDir,
  controlGrp,
  mergeExp1 = NA,
  mergeExp2 = NA,
  mergeExp3 = NA,
  mergeExp4 = NA,
  singleFish = FALSE,
  grporder = NA,
  skipNight0,
  avgDayNight = TRUE
)
}
\arguments{
\item{avgDayNight}{}
}
\description{
Title
}
