% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DvSorter.R
\name{dvSorter}
\alias{dvSorter}
\title{Sorts Raw DanioVision .txt Files}
\usage{
dvSorter(
  ffDir,
  zt0 = "09:00:00",
  date0 = NA,
  time0 = NA,
  dayduration = 14,
  pxwell = 8280.75
)
}
\arguments{
\item{pxwell}{}
}
\description{
Sorts Raw DanioVision .txt Files
}
