% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRAWs.R
\name{findLightTransitionFrame}
\alias{findLightTransitionFrame}
\title{Finds light transition frame}
\usage{
findLightTransitionFrame(Zeitgeberdurations, transitionHour)
}
\arguments{
\item{Zeitgeberdurations}{Series of timestamps in 'Zeitgeber durations', i.e. number of hours since (typically) 9 AM on day 0.}

\item{transitionHour}{Time in Zeitgeber duration of the light transition to search for.}
}
\value{
Frame index.
}
\description{
Finds frame index at which a given light transition occurred.
}
