% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{fingerprintSimilarity}
\alias{fingerprintSimilarity}
\title{Title}
\usage{
fingerprintSimilarity(
  fgp,
  metric = "mean",
  simScore,
  grporder = NA,
  removeControl = FALSE,
  controlGrp
)
}
\arguments{
\item{controlGrp}{}
}
\description{
Title
}
