% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTrace.R
\name{ggActivityTraceByGroup}
\alias{ggActivityTraceByGroup}
\title{Title}
\usage{
ggActivityTraceByGroup(
  ffpath,
  genopath,
  dayduration = 14,
  smoothOrNo = FALSE,
  smooth_nsecs = 30 * 60,
  binOrNo = TRUE,
  bin_nsecs = 10 * 60,
  ribbon = "sem",
  grporder = NA,
  onlyWell = NA,
  tracecols,
  tracedash = NA,
  ribboncols,
  linethick = 0.4,
  xname,
  yname,
  xtextOrNo = TRUE,
  ytextOrNo = TRUE,
  ymin = 0,
  ymax,
  xstart = 0,
  xstop = 0,
  trimstart = 0,
  trimstop = 0,
  xtick = 24,
  xmajorOrNo = TRUE,
  ymajorOrNo = TRUE,
  nightBgOrNo = FALSE,
  sunlinesOrNo = FALSE,
  markTimes = NA,
  legendOrNo = TRUE,
  exportOrNo = TRUE,
  exportPath,
  width = 75,
  height = 55
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
