% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTrace.R
\name{ggActivityTraceGrid}
\alias{ggActivityTraceGrid}
\title{Title}
\usage{
ggActivityTraceGrid(
  ffpath,
  genopath,
  dayduration = 14,
  smoothOrNo = FALSE,
  smooth_nsecs = 30 * 60,
  binOrNo = TRUE,
  bin_nsecs = 10 * 60,
  onlyWell = NA,
  tracecols = NA,
  tracedash = NA,
  linethick = 0.4,
  ymin = 0,
  ymax,
  xstart = 0,
  xstop = 0,
  trimstart = 0,
  trimstop = 0,
  xmajorOrNo = TRUE,
  ymajorOrNo = TRUE,
  sunlinesOrNo = FALSE,
  nightBgOrNo = FALSE,
  markTimes = NA,
  nrow,
  ncol,
  exportOrNo = TRUE,
  exportPath,
  width = 75,
  height = 55
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
