% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDeltaPx.R
\name{ggDeltaPx}
\alias{ggDeltaPx}
\title{Title}
\usage{
ggDeltaPx(
  ffpath,
  well,
  zhstart = NA,
  zhstop = NA,
  frstart = NA,
  frstop = NA,
  colour = NA,
  ymax = 10,
  xnsecs = 10,
  width = 100,
  height = 50,
  exportPath
)
}
\arguments{
\item{exportPath}{}
}
\description{
Title
}
