% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFingerprint.R
\name{ggFingerprintGrid}
\alias{ggFingerprintGrid}
\title{Title}
\usage{
ggFingerprintGrid(
  fgp,
  grporder = NA,
  onlyExp = NA,
  removeParam = NA,
  colours = NA,
  connectOrNo = TRUE,
  legendOrNo = TRUE,
  ynameOrNo = TRUE,
  ytextOrNo = TRUE,
  xtextOrNo = TRUE,
  xParamNum = FALSE,
  nightBgOrNo = TRUE,
  ymin,
  ymax,
  nrow,
  ncol,
  exportPath,
  width = 150,
  height = 100
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
