% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFramebyFrame.R
\name{ggFramerate}
\alias{ggFramerate}
\title{Title}
\usage{
ggFramerate(
  ffpath,
  zebpath,
  subsample = TRUE,
  subsample_by = 1000,
  xstart = 0,
  xstop = 0,
  ymin = 0,
  ymax = 30,
  sunlines = FALSE,
  dayduration = 14,
  xname = "hours since first 9 AM",
  yname = "frames-per-second",
  exportOrNo = FALSE,
  width = 75,
  height = 55,
  exportPath
)
}
\arguments{
\item{exportPath}{}
}
\description{
Title
}
