% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglegacyFingerprint.R
\name{gglegacyFingerprint}
\alias{gglegacyFingerprint}
\title{Title}
\usage{
gglegacyFingerprint(
  lFgp,
  onlyGrp = NA,
  colours = NA,
  legendOrNo = TRUE,
  ynameOrNo = TRUE,
  ytextOrNo = TRUE,
  xtextOrNo = TRUE,
  nightBgOrNo = TRUE,
  ymin,
  ymax,
  exportOrNo = FALSE,
  exportPath,
  width = 150,
  height = 100
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
