% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFingerprint.R
\name{ggPairwiseHeat}
\alias{ggPairwiseHeat}
\title{Title}
\usage{
ggPairwiseHeat(
  fgp,
  metric = "mean",
  simScore,
  grporder = NA,
  dbgorder = NA,
  removeControl = TRUE,
  controlGrp,
  medianMid = FALSE,
  minCol = NA,
  maxCol = NA,
  onlyHalf = NA,
  scoreSize = NA,
  legendOrNo = TRUE,
  labelsOrNo = TRUE,
  width = 250,
  height = 200,
  exportPath
)
}
\arguments{
\item{exportPath}{}
}
\description{
Title
}
