% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggParameter.R
\name{ggParameter}
\alias{ggParameter}
\title{Title}
\usage{
ggParameter(
  pa,
  grporder = NA,
  skipNight0 = FALSE,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA,
  sameClutch1 = NA,
  sameClutch2 = NA,
  sameClutch3 = NA,
  poolDayNight = FALSE,
  onlyExp = NA,
  onlyDayorNight = NA,
  onlyWin = NA,
  colours = NA,
  fainterExp = FALSE,
  faintMax = 0.5,
  ymin = NA,
  ymax = NA,
  dotSize = 0.5,
  connectMean = FALSE,
  legendOrNo = TRUE,
  xtextOrNo = TRUE,
  ynameOrNo = TRUE,
  yunitOrNo = FALSE,
  titleOrNo = TRUE,
  blankTitle = FALSE,
  nightBgOrNo = TRUE,
  statsOrNo = FALSE,
  silent = FALSE,
  width,
  height,
  exportPath
)
}
\arguments{
\item{connectMean}{}
}
\description{
Title
}
