% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggParameter.R
\name{ggParameterGrid}
\alias{ggParameterGrid}
\title{Title}
\usage{
ggParameterGrid(
  paDir,
  grporder = NA,
  skipNight0 = FALSE,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA,
  sameClutch1 = NA,
  sameClutch2 = NA,
  sameClutch3 = NA,
  poolDayNight = FALSE,
  onlyExp = NA,
  onlyDayorNight = NA,
  onlyWin = NA,
  colours = NA,
  fainterExp = FALSE,
  faintMax = 0.5,
  legendOrNo = TRUE,
  dotSize = 0.5,
  connectMean = FALSE,
  ynameOrNo = TRUE,
  yunitOrNo = FALSE,
  xtextOrNo = TRUE,
  titleOrNo = TRUE,
  nightBgOrNo = TRUE,
  keysOrNo = FALSE,
  statsOrNo = FALSE,
  ncol,
  nrow,
  width,
  height,
  exportPath
)
}
\arguments{
\item{connectMean}{}
}
\description{
Title
}
