% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTrace.R
\name{ggSleepGrid}
\alias{ggSleepGrid}
\title{Title}
\usage{
ggSleepGrid(
  zzz,
  genopath,
  boxnum = 1,
  grp,
  howMany = "all",
  smoothOrNo = FALSE,
  smooth_npoints,
  raiseBottom = FALSE,
  bottomborder,
  fill,
  xtext = FALSE,
  ytext = FALSE,
  xmajor = FALSE,
  ymajor = FALSE,
  xminor = FALSE,
  yminor = FALSE,
  xstart = 14,
  xstop = 24,
  ymin = 0,
  ymax = 8,
  grid_nrow = 4,
  grid_ncol = 2,
  spacing = 0.01,
  exportOrNo = TRUE,
  exportPath,
  width = 159.1,
  height = 246.1
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
