% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFramebyFrame.R
\name{ggSleepLatencyGrid}
\alias{ggSleepLatencyGrid}
\title{Title}
\usage{
ggSleepLatencyGrid(
  pa,
  grporder = NA,
  skipNight0 = FALSE,
  colours = NA,
  legendOrNo = TRUE,
  xtextOrNo = TRUE,
  xnameOrNo = TRUE,
  ynameOrNo = TRUE,
  nightBgOrNo = TRUE,
  xmaxh = NA,
  detailsOrNo = TRUE,
  exportDir,
  width = 150,
  height = 100,
  dayduration = 14
)
}
\arguments{
\item{dayduration}{}
}
\description{
Title
}
