% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFramebyFrame.R
\name{ggSleepLatencySurvival}
\alias{ggSleepLatencySurvival}
\title{Title}
\usage{
ggSleepLatencySurvival(
  pa,
  grporder = NA,
  skipNight0 = FALSE,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA,
  onlyDayorNight = NA,
  onlyWin = NA,
  colours = NA,
  legendOrNo = TRUE,
  xtextOrNo = TRUE,
  xnameOrNo = TRUE,
  ynameOrNo = TRUE,
  nightBgOrNo = TRUE,
  xmaxh = NA,
  detailsOrNo = TRUE,
  exportOrNo = TRUE,
  exportDir,
  width = 120,
  height = 70,
  dayduration = 14,
  woi = NA
)
}
\arguments{
\item{woi}{}
}
\description{
Title
}
