% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTrace.R
\name{ggTrace}
\alias{ggTrace}
\title{Title}
\usage{
ggTrace(
  tc,
  genopath,
  dayduration,
  ribbon,
  grporder,
  onlyWell,
  tracecols,
  tracedash,
  ribboncols,
  linethick,
  xname,
  yname,
  xtextOrNo,
  ytextOrNo,
  ymin,
  ymax,
  xstart,
  xstop,
  trimstart,
  trimstop,
  xtick = xtick,
  xmajorOrNo,
  ymajorOrNo,
  nightBgOrNo,
  sunlinesOrNo,
  markTimes,
  legendOrNo,
  gridMode = FALSE,
  exportOrNo,
  exportPath,
  width,
  height
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
