% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBhvParams.R
\name{importBhvParams}
\alias{importBhvParams}
\title{Title}
\usage{
importBhvParams(
  paDir,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA,
  grporder = NA,
  skipNight0 = FALSE
)
}
\arguments{
\item{skipNight0}{}
}
\description{
Title
}
