% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacyFingerprint.R
\name{legacyFingerprint}
\alias{legacyFingerprint}
\title{Title}
\usage{
legacyFingerprint(matPath, conGrp, treGrp, nights = c(2, 3), days = c(2, 3))
}
\arguments{
\item{days}{}
}
\description{
Title
}
