% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEstats.R
\name{LMEdaynight}
\alias{LMEdaynight}
\title{Title}
\usage{
LMEdaynight(
  pa,
  grporder = NA,
  skipNight0 = FALSE,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA,
  sameClutch1 = NA,
  sameClutch2 = NA,
  sameClutch3 = NA,
  silent = FALSE,
  detailsOrNo = TRUE
)
}
\arguments{
\item{detailsOrNo}{}
}
\description{
Title
}
