% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEstats.R
\name{LMEreport}
\alias{LMEreport}
\title{Title}
\usage{
LMEreport(
  paDir,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA,
  sameClutch1 = NA,
  sameClutch2 = NA,
  sameClutch3 = NA,
  grporder = NA,
  skipNight0 = FALSE,
  silent = FALSE,
  detailsOrNo = FALSE,
  exportPath
)
}
\arguments{
\item{exportPath}{}
}
\description{
Title
}
