% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviourParameter.R
\name{multiBehaviourParameter}
\alias{multiBehaviourParameter}
\title{Title}
\usage{
multiBehaviourParameter(
  parameters,
  ffpath,
  genopath,
  woi = NA,
  skipNight0 = FALSE,
  zthr_min = 1,
  inaThr = 0,
  dayduration = 14
)
}
\arguments{
\item{inaThr}{}
}
\description{
Title
}
