% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviourParameter.R
\name{paramReadPivot}
\alias{paramReadPivot}
\title{Title}
\usage{
paramReadPivot(
  pa,
  grporder = NA,
  skipNight0,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA
)
}
\arguments{
\item{poolExp3}{}
}
\description{
Title
}
