% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviourParameter.R
\name{paramSummary}
\alias{paramSummary}
\title{Title}
\usage{
paramSummary(
  pa,
  skipNight0 = FALSE,
  poolExp1 = NA,
  poolExp2 = NA,
  poolExp3 = NA,
  poolExps = FALSE,
  poolWins = FALSE
)
}
\arguments{
\item{poolWins}{}
}
\description{
Title
}
