% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePulses.R
\name{replacePulses}
\alias{replacePulses}
\title{Title}
\usage{
replacePulses(
  ffpath,
  sentinel,
  genopath = NA,
  onlyGrp = NA,
  zhstartsd,
  zhstopsd,
  px_thr = 2,
  nframes_thr = 2,
  extendBefore = 2,
  extendAfter = 2,
  replaceBy = 0
)
}
\arguments{
\item{replaceBy}{}
}
\description{
Title
}
