% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepParameter.R
\name{sleepParameter}
\alias{sleepParameter}
\title{Title}
\usage{
sleepParameter(
  dn,
  parameter,
  zebpath,
  woi = woi,
  zthr_min,
  inaThr,
  dayduration
)
}
\arguments{
\item{dayduration}{}
}
\description{
Title
}
