% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VpSorter.R
\name{SpeaknRecord}
\alias{SpeaknRecord}
\title{Speak and Record to Log File}
\usage{
SpeaknRecord(..., header = 0, warning = FALSE)
}
\arguments{
\item{warning}{}
}
\description{
Speak and Record to Log File
}
