% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryActivityCourse.R
\name{summaryActivityCourse}
\alias{summaryActivityCourse}
\title{Title}
\usage{
summaryActivityCourse(
  ffpath,
  genopath,
  smoothOrNo = TRUE,
  smooth_nsecs = 30 * 60,
  binOrNo = TRUE,
  bin_nsecs = 10 * 60
)
}
\arguments{
\item{bin_nsecs}{}
}
\description{
Title
}
