% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepFromFrames.R
\name{summarySleepCourse}
\alias{summarySleepCourse}
\title{Title}
\usage{
summarySleepCourse(
  ffpath,
  genopath,
  zthr_min,
  inaThr,
  epo_min = 10,
  smoothOrNo = FALSE,
  smooth_npoints
)
}
\arguments{
\item{smooth_npoints}{}
}
\description{
Title
}
