% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeboutParameter_onefish.R
\name{transitionsToActiveBoutStartStop}
\alias{transitionsToActiveBoutStartStop}
\title{Title}
\usage{
transitionsToActiveBoutStartStop(colAI)
}
\arguments{
\item{colAI}{}
}
\description{
Title
}
