% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VpSorter.R
\name{vpSorter}
\alias{vpSorter}
\title{Sorts RAW .xls Files}
\usage{
vpSorter(
  ffDir,
  zebpath = NA,
  boxGen = 2,
  twoBoxMode = TRUE,
  boxnum = 1,
  zt0 = "09:00:00",
  date0 = NA,
  time0 = NA,
  dayduration = 14,
  exportXlsOrNo = FALSE
)
}
\arguments{
\item{exportXlsOrNo}{}
}
\description{
Sorts RAW .xls Files
}
