% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviourParameter.R
\name{behaviourParameter}
\alias{behaviourParameter}
\title{Title}
\usage{
behaviourParameter(
  parameter,
  ffpath,
  genopath,
  woi = NA,
  zthr_min = 1,
  inaThr = 0
)
}
\arguments{
\item{inaThr}{}
}
\description{
Title
}
