% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFramebyFrame.R
\name{ggFingerprint}
\alias{ggFingerprint}
\title{Title}
\usage{
ggFingerprint(
  fgp,
  onlyFish = NA,
  metric = "mean",
  grporder = NA,
  controlGrp,
  removeControl = FALSE,
  onlyExp = NA,
  removeParam = NA,
  colours = NA,
  connectOrNo = TRUE,
  legendOrNo = TRUE,
  ynameOrNo = TRUE,
  ytextOrNo = TRUE,
  xtextOrNo = TRUE,
  xParamNum = FALSE,
  nightBgOrNo = TRUE,
  ymin,
  ymax,
  exportOrNo = TRUE,
  exportPath,
  width = 150,
  height = 100
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
